功能：添加本地商品主页 UI 资源管理 - 引入了一个新的 LocalGoodsHomeUiAssets 组件，用于管理本地商品主页资产的显示。 - 更新了路由器，新增了本地商品主页 UI 资产的路由。 - 优化了 LocalGoodsHomeConfig 视图，使其能够显示产品图片并具备备用处理功能。 - 增加了图片上传功能，并改进了资产管理特性，以提升用户体验。

+++ backend/src/models/MarketFavoriteItem.js
@@ -10,7 +10,7 @@ module.exports = (sequelize, DataTypes) => {
   }
   MarketFavoriteItem.init(
     {
-      user_id: { type: DataTypes.INTEGER, allowNull: false },
+      user_id: { type: DataTypes.BIGINT, allowNull: false },
       goods_id: { type: DataTypes.INTEGER, allowNull: false },
       shop_id: { type: DataTypes.INTEGER, allowNull: false }
     },
